Renamed settings and changed settings component.
